Refactoring mobilization's domain configuration form into a reusable component #531
