+++ client/mobilizations/components/index.js
@@ -2,3 +2,8 @@ export { default as MobilizationBasicsForm } from './mobilization-basics-form'
 export { default as SettingsMenu } from './settings-menu'
 export { default as PageTabLayout } from './page-tab-layout'
 export { default as PageHeader } from './page-header'
+
+//
+// Components that implements FormRedux component
+//
+export { default as FormDomain } from './form-domain'
