+++ client/ux/components/flat-form/flat-form.scss
@@ -3,3 +3,8 @@
   background-color: #fff;
   padding: 2rem 2rem 0;
 }
+
+.ux--flat-form .form-group {
+  margin-left: -2rem !important;
+  margin-right: -2rem !important;
+}
